requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.37
sqlmodel==0.0.21
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
pydantic-settings==2.6.1
python-multipart==0.0.9